mcp/requirements.txt
fastmcp==3.1.1
python-dotenv==1.2.2
starlette==0.52.1
uvicorn==0.42.0
httpx==0.28.1
pydantic==2.12.5
pydantic-settings==2.13.1